refactor: Create TeacherCurriculum as entrypoint for teacher's courses
